refactor to make Base card more generic
